+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Ability to add/change hyperlinks anchored to bookmarks/headings
 * The following functions are added: GetRows, GetCols, GetCount, GetHidden, SetHidden, GetColumnWidth, SetColumnWidth, GetRowHeight, SetRowHeight
 * The following properties are added: Rows, Cols, Count, Hidden, ColumnWidth, Width, RowHeight, Height
 * Fix change active cell in selection across merge. Previously, passing through the first cell of the merge range, we fell into the merge range, even if it was not selected (through the selection of a row / column)
+* Fix selection when selecting row/col/all
 
 ### Presentation Editor
 *